+++ lib/neatogen/randomkit.h
@@ -86,11 +86,6 @@ extern void rk_seed(unsigned long seed, rk_state *state);
 extern unsigned long rk_random(rk_state *state);
 
 /*
- * Returns a random unsigned long between 0 and ULONG_MAX inclusive
- */
-extern unsigned long rk_ulong(rk_state *state);
-
-/*
  * Returns a random unsigned long between 0 and max inclusive.
  */
 extern unsigned long rk_interval(unsigned long max, rk_state *state);