+++ lib/neatogen/sgd.c
@@ -23,13 +23,10 @@ static float calculate_stress(float *pos, term_sgd *terms, int n_terms) {
     return stress;
 }
 // it is much faster to shuffle term rather than pointers to term, even though the swap is more expensive
-static rk_state rstate;
-static void fisheryates_shuffle(term_sgd *terms, int n_terms) {
+static void fisheryates_shuffle(term_sgd *terms, int n_terms, rk_state *rstate) {
     int i;
     for (i=n_terms-1; i>=1; i--) {
-        // srand48() is called in neatoinit.c, so no need to seed here
-        //int j = (int)(drand48() * (i+1));
-        int j = rk_interval(i, &rstate);
+        int j = rk_interval(i, rstate);
 
         term_sgd temp = terms[i];
         terms[i] = terms[j];
@@ -188,10 +185,8 @@ void sgd(graph_t *G, /* input graph */
     float w_min = terms[0].w, w_max = terms[0].w;
     int ij;
     for (ij=1; ij<n_terms; ij++) {
-        if (terms[ij].w < w_min)
-            w_min = terms[ij].w;
-        if (terms[ij].w > w_max)
-            w_max = terms[ij].w;
+        w_min = fminf(w_min, terms[ij].w);
+        w_max = fmaxf(w_max, terms[ij].w);
     }
     // note: Epsilon is different from MODE_KK and MODE_MAJOR as it is a minimum step size rather than energy threshold
     //       MaxIter is also different as it is a fixed number of iterations rather than a maximum
@@ -217,9 +212,10 @@ void sgd(graph_t *G, /* input graph */
         start_timer();
     }
     int t;
+    rk_state rstate;
     rk_seed(0, &rstate); // TODO: get seed from graph
     for (t=0; t<MaxIter; t++) {
-        fisheryates_shuffle(terms, n_terms);
+        fisheryates_shuffle(terms, n_terms, &rstate);
         float eta = eta_max * exp(-lambda * t);
         for (ij=0; ij<n_terms; ij++) {
             // cap step size